Time: 1304 ms (100%), Space: 75.8 MB (100%) - LeetHub

+++ 4044-count-good-cyclic-rotations/4044-count-good-cyclic-rotations.java
@@ -1,8 +1,8 @@
 class Solution {
     public int countGoodRotations(int[] nums) {
       
-       int total=0;
-       int leftsum=0;
+       long total=0;
+       long leftsum=0;
        int count=0;
 
        for(int i=0;i<nums.length;i++){
@@ -13,7 +13,7 @@ class Solution {
            total=total+nums[i];
        } 
       
-      int rigthsum=0;
+      long rigthsum=0;
       rigthsum=total-leftsum;
       if(leftsum!=rigthsum){
             count++;
